Move tests, remove `check_kid`
