+++ identity_jose/src/jwu/base64.rs
@@ -8,14 +8,17 @@ use serde_json::to_vec;
 
 use crate::error::Result;
 
+/// Encode the given slice in url-safe base64.
 pub fn encode_b64(data: impl AsRef<[u8]>) -> String {
   base64::encode_config(data.as_ref(), base64::URL_SAFE_NO_PAD)
 }
 
+/// Decode the given url-safe base64-encoded slice into its raw bytes.
 pub fn decode_b64(data: impl AsRef<[u8]>) -> Result<Vec<u8>> {
   base64::decode_config(data.as_ref(), base64::URL_SAFE_NO_PAD).map_err(Into::into)
 }
 
+/// Serialize the given data into JSON and encode the result in url-safe base64.
 pub fn encode_b64_json<T>(data: &T) -> Result<String>
 where
   T: Serialize,
@@ -23,6 +26,7 @@ where
   to_vec(data).map(encode_b64).map_err(Into::into)
 }
 
+/// Decode the given url-safe base64-encoded slice into its raw bytes and try to deserialize it into `T`.
 pub fn decode_b64_json<T>(data: impl AsRef<[u8]>) -> Result<T>
 where
   T: DeserializeOwned,
